requirements.txt
scipy==1.6.2
numpy==1.19.2
pandas==1.2.3
tensorflow==2.3.0
mysql_connector_repackaged==0.3.1
